+++ src/com/mitp0sh/jaclaff/attributes/AttributeCustom.java
@@ -22,8 +22,8 @@ private byte[] content = new byte[0];
     {
 		AttributeCustom attribute = new AttributeCustom();			
 		
-		byte[] content = new byte[attribute.getAttributeLength()];		
-		dis.read(content, 0, attribute.getAttributeLength());		
+		byte[] content = new byte[(int)attribute.getAttributeLength()];		
+		dis.read(content, 0, (int)attribute.getAttributeLength());		
 		attribute.setContent(content);
 		
 		return attribute;
