+++ src/com/mitp0sh/jaclaff/attributes/localvariabletable/LocalVariableTableEntry.java
@@ -9,61 +9,61 @@ import com.mitp0sh.jaclaff.util.PNC;
 
 public class LocalVariableTableEntry 
 {
-	private short              start_pc = 0;
-	private short                length = 0;
-	private short            name_index = 0;
-	private short       signature_index = 0;
-	private short                 index = 0;
+	private int              start_pc = 0;
+	private int                length = 0;
+	private int            name_index = 0;
+	private int       signature_index = 0;
+	private int                 index = 0;
 	
 	/* additional info */
 	private int instructionIndexStartPc = 0;
 	
-	public short getStart_pc() 
+	public int getStart_pc() 
 	{
 		return start_pc;
 	}
 
-	public void setStart_pc(short startPc) 
+	public void setStart_pc(int startPc) 
 	{
 		start_pc = startPc;
 	}
 
-	public short getLength() 
+	public int getLength() 
 	{
 		return length;
 	}
 
-	public void setLength(short length) 
+	public void setLength(int length) 
 	{
 		this.length = length;
 	}
 
-	public short getName_index() 
+	public int getName_index() 
 	{
 		return name_index;
 	}
 
-	public void setName_index(short nameIndex) 
+	public void setName_index(int nameIndex) 
 	{
 		name_index = nameIndex;
 	}
 
-	public short getSignature_index() 
+	public int getSignature_index() 
 	{
 		return signature_index;
 	}
 
-	public void setSignature_index(short signatureIndex) 
+	public void setSignature_index(int signatureIndex) 
 	{
 		signature_index = signatureIndex;
 	}
 	
-	public short getIndex() 
+	public int getIndex() 
 	{
 		return index;
 	}
 	
-	public void setIndex(short index) 
+	public void setIndex(int index) 
 	{
 		this.index = index;
 	}
@@ -82,11 +82,11 @@ public class LocalVariableTableEntry
     {
 		LocalVariableTableEntry localVariableTableEntry = new LocalVariableTableEntry();
 		
-		localVariableTableEntry.setStart_pc((short)dis.readUnsignedShort());
-		localVariableTableEntry.setLength((short)dis.readUnsignedShort());
-		localVariableTableEntry.setName_index((short)dis.readUnsignedShort());
-		localVariableTableEntry.setSignature_index((short)dis.readUnsignedShort());
-		localVariableTableEntry.setIndex((short)dis.readUnsignedShort());
+		localVariableTableEntry.setStart_pc(dis.readUnsignedShort());
+		localVariableTableEntry.setLength(dis.readUnsignedShort());
+		localVariableTableEntry.setName_index(dis.readUnsignedShort());
+		localVariableTableEntry.setSignature_index(dis.readUnsignedShort());
+		localVariableTableEntry.setIndex(dis.readUnsignedShort());
 	    
 		return localVariableTableEntry;
     }
